DEV | developer class get_file and update_file. Create folders and clean coding
