+++ .gitignore
@@ -15,3 +15,4 @@ venv.bak/
 __pycache__/
 *.py[cod]
 *$py.class
+__pycache__
