--- bucketuser_accessKeys.csv
@@ -1,2 +1,0 @@
-Access key ID,Secret access key
-AKIA4BSTJ23SAS2G2JPR,YznfNr9G73CRcjffC0zziJPfdmGZlcyhZsLrzrYV
