--- input/hello.txt
@@ -1,2 +1,0 @@
-123 123 123 123
-123 123 123 123
