--- main.py
@@ -1,4 +1,0 @@
-from upload_file import *
-
-AWS_ACCESS_KEY_ID = 'AKIA4BSTJ23SAS2G2JPR'
-AWS_SECRET_ACCESS_KEY = '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'
