--- read_buckets.py
@@ -1,8 +1,0 @@
-import boto3  # pip install boto3
-
-# Let's use Amazon S3
-s3 = boto3.resource("s3")
-
-# Print out bucket names
-for bucket in s3.buckets.all():
-    print(bucket.name)
