--- tmp/s2.csv
@@ -1,1 +1,0 @@
-assa afasf aff
